Improving gs_font_procs interface functions glyph_info and glyph_outline.
DETAILS:

The old interface definition provided insufficient data for pdfwrite.
Several hacks was used in pdfwrite to recover the luck of data.
This patch slightly enhances the interface and removes the hacks.
See new comments in gxfont.h about the enhancement.

The interface could be more complete - see comments in gxfont.h .
Now we do not apply efforts for refinishing it, because
there is no practical need and we have no tests for it.

This patch improves glyph positions with pdfwrite for fonts with
vertical CMaps. comparefiles do not test this case.

Minor change : fixed compiler warnings in gdevpdtd.c .

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3625 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/gdevpdtd.c
@@ -377,15 +377,15 @@ pdf_compute_font_descriptor(pdf_font_descriptor_t *pfd)
 	}
 	rect_merge(desc.FontBBox, info.bbox);
 	if (!info.num_pieces)
-	    desc.Ascent = max(desc.Ascent, info.bbox.q.y);
+	    desc.Ascent = max(desc.Ascent, (int)info.bbox.q.y);
 	if (notdef == GS_NO_GLYPH && gs_font_glyph_is_notdef(bfont, glyph)) {
 	    notdef = glyph;
-	    desc.MissingWidth = info.width[wmode].x;
+	    desc.MissingWidth = (int)info.width[wmode].x;
 	}
 	if (info.width[wmode].y != 0)
 	    fixed_width = min_int;
 	else if (fixed_width == 0)
-	    fixed_width = info.width[wmode].x;
+	    fixed_width = (int)info.width[wmode].x;
 	else if (info.width[wmode].x != fixed_width)
 	    fixed_width = min_int;
 	if (desc.Flags & FONT_IS_SYMBOLIC)
